requirements.txt
requests~=2.32.4
PyYAML~=5.4.1
lxml~=4.6.3